クラスを定義するファイルと、クラスを実行するファイルを分割。
